Create energia_code_stepper_motor_on_face_detection
The following code runs the stepper motor connected to the TIVA board only when a face is detected, this happens when a particular character is written on the serial port upon face detection, when the com port senses the character written on it, it instructs the TIVA board to run the motor. For the processing code for face detection look into the processing repository.